refactor: move `shouldRunCheck` to its own package
