+++ deps/should-run-check/README.md
@@ -1,0 +1,17 @@
+# @pnpm/deps.should-run-check
+
+> Determine whether a deps check can run
+
+<!--@shields('npm')-->
+[![npm version](https://img.shields.io/npm/v/@pnpm/deps.should-run-check.svg)](https://www.npmjs.com/package/@pnpm/deps.should-run-check)
+<!--/@-->
+
+## Installation
+
+```sh
+pnpm add @pnpm/deps.should-run-check
+```
+
+## License
+
+MIT
