+++ deps/status/src/index.ts
@@ -1,2 +1,1 @@
 export { type CheckDepsStatusOptions, checkDepsStatus } from './checkDepsStatus'
-export { type ShouldRunCheckEnv, DISABLE_DEPS_CHECK_ENV, SKIP_ENV_KEY, shouldRunCheck } from './shouldRunCheck'
