+++ deps/status/tsconfig.json
@@ -59,6 +59,9 @@
     },
     {
       "path": "../../workspace/state"
+    },
+    {
+      "path": "../should-run-check"
     }
   ]
 }
