+++ exec/lifecycle/package.json
@@ -36,7 +36,7 @@
   },
   "dependencies": {
     "@pnpm/core-loggers": "workspace:*",
-    "@pnpm/deps.status": "workspace:*",
+    "@pnpm/deps.should-run-check": "workspace:*",
     "@pnpm/directory-fetcher": "workspace:*",
     "@pnpm/error": "workspace:*",
     "@pnpm/link-bins": "workspace:*",
