+++ exec/plugin-commands-script-runners/src/exec.ts
@@ -3,7 +3,8 @@ import { docsUrl, type RecursiveSummary, throwOnCommandFail, readProjectManifest
 import { type LifecycleMessage, lifecycleLogger } from '@pnpm/core-loggers'
 import { FILTERING, UNIVERSAL_OPTIONS } from '@pnpm/common-cli-options-help'
 import { type Config, types } from '@pnpm/config'
-import { type CheckDepsStatusOptions, DISABLE_DEPS_CHECK_ENV, SKIP_ENV_KEY } from '@pnpm/deps.status'
+import { DISABLE_DEPS_CHECK_ENV, SKIP_ENV_KEY } from '@pnpm/deps.should-run-check'
+import { type CheckDepsStatusOptions } from '@pnpm/deps.status'
 import { makeNodeRequireOption } from '@pnpm/lifecycle'
 import { logger } from '@pnpm/logger'
 import { tryReadProjectManifest } from '@pnpm/read-project-manifest'
