+++ exec/plugin-commands-script-runners/tsconfig.json
@@ -34,6 +34,9 @@
       "path": "../../crypto/hash"
     },
     {
+      "path": "../../deps/should-run-check"
+    },
+    {
       "path": "../../deps/status"
     },
     {
